Coverage testing: lower branch to 80%
Considering the very small number of branches in this project, the coverage requirement of 90% can be misleading. Even one non-covered branch pretty much drops it under 90%.

The one uncovered branch was required for similar use case when creating slugs for this dictionary. It is unclear if slugify has changed, but before there were slugs starting with dash.